feat: :sparkles: implement uploader reducer and context
